requirements.txt
numpy
asyncio
websockets
mysql-connector